# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DreamCo Command Center API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type BotStatus = (typeof BotStatus)[keyof typeof BotStatus];

export const BotStatus = {
  active: "active",
  idle: "idle",
  error: "error",
  unknown: "unknown",
} as const;

export type BotTier = (typeof BotTier)[keyof typeof BotTier];

export const BotTier = {
  FREE: "FREE",
  PRO: "PRO",
  ENTERPRISE: "ENTERPRISE",
} as const;

export interface Bot {
  name: string;
  repoPath?: string;
  status: BotStatus;
  tier: BotTier;
  category: string;
  description?: string;
  /** @nullable */
  lastHeartbeat?: string | null;
  pendingPRs?: number;
  revenue?: number;
  /** @nullable */
  lastUpdate?: string | null;
}

export interface Repo {
  name: string;
  fullName: string;
  /** @nullable */
  description: string | null;
  isPrivate: boolean;
  url: string;
  stars: number;
  forks: number;
  updatedAt: string;
  /** @nullable */
  language?: string | null;
  openIssues?: number;
}

export interface Commit {
  sha: string;
  message: string;
  author: string;
  date: string;
  repoName: string;
  url: string;
}

export type RepoActivityRecentEventsItem = {
  type: string;
  repo: string;
  description: string;
  date: string;
};

export interface RepoActivity {
  totalCommits: number;
  totalRepos: number;
  totalBots: number;
  activeRepos: number;
  recentEvents: RepoActivityRecentEventsItem[];
}

export interface RevenueMetrics {
  totalRevenue: number;
  dailyRevenue: number;
  weeklyRevenue: number;
  monthlyRevenue: number;
  dailyTarget: number;
  weeklyTarget: number;
  monthlyTarget: number;
  currency: string;
  connected?: boolean;
}

export interface Subscription {
  id: string;
  customer: string;
  plan: string;
  status: string;
  amount: number;
  currency: string;
  currentPeriodEnd: string;
}

export interface Transaction {
  id: string;
  amount: number;
  currency: string;
  status: string;
  description: string;
  date: string;
}

export interface BuddyMessageInput {
  message: string;
  sessionId?: string;
}

export interface BuddyResponse {
  message: string;
  sessionId: string;
  timestamp: string;
  /** @nullable */
  emotion?: string | null;
}

export type ChatMessageRole =
  (typeof ChatMessageRole)[keyof typeof ChatMessageRole];

export const ChatMessageRole = {
  user: "user",
  buddy: "buddy",
} as const;

export interface ChatMessage {
  id: string;
  role: ChatMessageRole;
  content: string;
  timestamp: string;
  /** @nullable */
  emotion?: string | null;
}

export interface DashboardSummary {
  totalBots: number;
  activeBots: number;
  idleBots: number;
  totalRepos: number;
  dailyRevenue: number;
  monthlyRevenue: number;
  dailyTarget: number;
  monthlyTarget: number;
  totalSubscriptions: number;
  recentCommits: number;
  profitTargetDaily?: number;
  profitTargetMonthly?: number;
}

export interface Tier {
  name: string;
  priceMonthly: number;
  requestsPerMonth: number;
  concurrentBots: number;
  models: string[];
}