# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: DreamCo Command Center API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bots/{name}:
    get:
      operationId: getBot
      tags: [bots]
      summary: Get bot details
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bot details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"
        "404":
          description: Bot not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /buddy/chat:
    post:
      operationId: sendBuddyMessage
      tags: [buddy]
      summary: Send a message to BuddyBot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BuddyMessageInput"
      responses:
        "200":
          description: BuddyBot response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BuddyResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    Bot:
      type: object
      required: [name, status, tier, category]
      properties:
        name:
          type: string
        repoPath:
          type: string
        status:
          type: string
          enum: [active, idle, error, unknown]
        tier:
          type: string
          enum: [FREE, PRO, ENTERPRISE]
        category:
          type: string
        description:
          type: string
        lastHeartbeat:
          type: ["string", "null"]
        pendingPRs:
          type: integer
        revenue:
          type: number
        lastUpdate:
          type: ["string", "null"]
...
    BuddyMessageInput:
      type: object
      required: [message]
      properties:
        message:
          type: string
        sessionId:
          type: string
    BuddyResponse:
      type: object
      required: [message, sessionId, timestamp]
      properties:
        message:
          type: string
        sessionId:
          type: string
        timestamp:
          type: string
        emotion:
          type: ["string", "null"]
...