# lib/api-zod/src/generated/types/bot.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DreamCo Command Center API
 * OpenAPI spec version: 0.1.0
 */
import type { BotStatus } from "./botStatus";
import type { BotTier } from "./botTier";

export interface Bot {
  name: string;
  repoPath?: string;
  status: BotStatus;
  tier: BotTier;
  category: string;
  description?: string;
  /** @nullable */
  lastHeartbeat?: string | null;
  pendingPRs?: number;
  revenue?: number;
  /** @nullable */
  lastUpdate?: string | null;
}